real real last at home

+++ internal/model/database/user.go
@@ -1,7 +1,7 @@
 package database
 
 type User struct {
-	ID       int    `gorm:"primary_key" json:"id"`
+	ID       int64  `gorm:"primary_key" json:"id"`
 	Username string `gorm:"type:varchar(20);not null" binding:"required"`
 	Password string
 	Email    string
